feat: don't allow modification if answers are locked and display a notification (#210)
* feat: don't allow modification if answers are locked

* refactor: update question page to use questions locked store

* fix: update electionCanEditQuestions ACL

* feat(acl): don't allow updating candidate basic info when election is locked

* fix: documentation indentation

* feat: restructure some elements in basic info

* fix: Joonatan's complaints

---------

+++ frontend/src/routes/[[lang=locale]]/candidate/(protected)/profile/+page.svelte
@@ -15,6 +15,8 @@
   import {getContext} from 'svelte';
   import type {StrapiGenderData, StrapiLanguageData} from '$lib/api/getData.type';
   import type {Language} from '$lib/types/candidateAttributes';
+  import Warning from '$lib/components/warning/Warning.svelte';
+  import InputContainer from './InputContainer.svelte';
   import type {CandidateContext} from '$lib/utils/candidateStore';
 
   const basicInfoFields = ['firstName', 'lastName', 'party'];
@@ -23,20 +25,29 @@
     'pointer-events-none label-sm whitespace-nowrap label mx-6 my-2 text-secondary';
   const disclaimerClass = 'mx-6 my-0 p-0 text-sm text-secondary';
   const headerClass = 'uppercase mx-6 my-0 p-0 small-label';
-  const selectClass = 'select select-sm w-full text-right text-primary';
+  const selectClass =
+    'select select-sm w-full text-right text-primary disabled:bg-base-100 disabled:border-none';
   const inputClass =
     'input-ghost flex justify-end text-right input input-sm w-full pr-2 disabled:border-none disabled:bg-base-100';
   const iconClass = 'text-secondary my-auto flex-shrink-0';
   const buttonContainerClass = 'pr-6';
-  const inputContainerClass = 'flex w-full pr-6';
 
   // get the user from authContext
-  const {userStore, loadUserData} = getContext<CandidateContext>('candidate');
+  const {
+    userStore,
+    questionsLockedStore,
+    opinionQuestionsFilledStore,
+    basicInfoFilledStore,
+    loadUserData
+  } = getContext<CandidateContext>('candidate');
   const user = get(userStore);
+  $: questionsLocked = $questionsLockedStore;
+  $: opinionQuestionsFilled = $opinionQuestionsFilledStore;
+  $: basicInfoFilled = $basicInfoFilledStore;
 
   let loading = false;
 
-  let {gender, motherTongues, birthday, photo, unaffiliated, manifesto, nominations} = {
+  let {gender, motherTongues, birthday, photo, unaffiliated, manifesto, nomination} = {
     gender: {
       id: undefined
     },
@@ -62,7 +73,7 @@
         birthday,
         photo,
         unaffiliated,
-        nominations,
+        nomination,
         manifesto: Object.values(manifesto).join('')
       });
     previousStateHash = previousStateHash ?? getCurrentHash();
@@ -85,25 +96,33 @@
   let uploadPhoto: () => Promise<void>;
 
   const submitForm = async () => {
+    if (questionsLocked) {
+      await goto($getRoute(Route.CandAppHome));
+      return;
+    }
     loading = true;
     dirty = false;
     try {
-      await uploadPhoto();
-      await updateBasicInfo(
-        manifesto,
-        birthday,
-        genderID ?? undefined,
-        photo,
-        unaffiliated,
-        motherTongues
-      );
+      if (!questionsLocked) {
+        await uploadPhoto();
+        await updateBasicInfo(
+          manifesto,
+          birthday,
+          genderID ?? undefined,
+          photo,
+          unaffiliated,
+          motherTongues
+        );
+      }
 
       // Update the database-saved manifesto in order to detect changes
       savedManifesto = manifesto;
       manifestoTextArea.deleteLocal();
 
       await loadUserData(); // reload user data so it's up to date
-      await goto($getRoute(Route.CandAppQuestions));
+      if (!opinionQuestionsFilled && !questionsLocked)
+        await goto($getRoute(Route.CandAppQuestions));
+      else await goto($getRoute(Route.CandAppHome));
     } catch (error) {
       errorMessage = $t('candidateApp.basicInfo.errorMessage');
     } finally {
@@ -113,22 +132,6 @@
 
   // the dot symbol for separating info string
   const dot = '\u22C5';
-
-  // map nominations into objects
-  const nominationFields = nominations?.map((nom) => {
-    const constituency = translate(nom.constituency?.name);
-    const party = translate(nom.party?.shortName);
-    const electionSymbol = nom.electionSymbol;
-    return {
-      nominationID: nom.id.toString(),
-      constituency,
-      party,
-      electionSymbol,
-      fieldText: `${constituency} ${dot} ${party} ${
-        electionSymbol ? dot + ' ' + electionSymbol : ''
-      }`
-    };
-  });
 
   // basic information
   const basicInfoData: Record<string, string | number | undefined> = {
@@ -172,11 +175,25 @@
     }
   };
 
+  let submitButtonText = '';
+  $: {
+    if (!opinionQuestionsFilled && !questionsLocked)
+      submitButtonText = $t('candidateApp.basicInfo.saveAndContinue');
+    else if (questionsLocked) submitButtonText = $t('candidateApp.basicInfo.return');
+    else submitButtonText = $t('candidateApp.basicInfo.saveAndReturn');
+  }
   const birthdayMin = '1800-01-01';
   const birthdayMax = new Date().toISOString().split('T')[0];
 </script>
 
 <BasicPage title={$t('candidateApp.basicInfo.title')} mainClass="bg-base-200">
+  <Warning display={!!questionsLocked} slot="note">
+    <p>{$t('candidateApp.homePage.editingNotAllowedNote')}</p>
+    {#if !opinionQuestionsFilled || !basicInfoFilled}
+      <p>{$t('candidateApp.homePage.editingNotAllowedPartiallyFilled')}</p>
+    {/if}
+  </Warning>
+
   <PreventNavigation active={dirty && !loading} />
   <form on:submit|preventDefault={submitForm}>
     <div class="flex flex-col items-center gap-16">
@@ -190,15 +207,14 @@
             <label for={field} class={labelClass}>
               {$t(`candidateApp.basicInfo.fields.${field}`)}
             </label>
-            <div class={inputContainerClass}>
+            <InputContainer locked>
               <input
                 type="text"
                 disabled
                 id={field}
                 value={basicInfoData[field]}
                 class={inputClass} />
-              <Icon name="locked" class={iconClass} />
-            </div>
+            </InputContainer>
           </Field>
         {/each}
         <p class={disclaimerClass} slot="footer">
@@ -210,21 +226,23 @@
         <p class={headerClass} slot="header">
           {$t('candidateApp.basicInfo.nominations')}
         </p>
-
-        {#each nominationFields ?? [] as nomination}
+        {#if nomination}
           <Field>
-            <label for={nomination.nominationID} class={labelClass}>{nomination.fieldText}</label>
-            <div class={inputContainerClass}>
+            <label for="nomination" class={labelClass}
+              >{`${translate(nomination.constituency?.shortName)} ${dot} ${translate(nomination.party.shortName)} ${
+                nomination.electionSymbol ? dot + ' ' + nomination.electionSymbol : ''
+              }`}</label>
+            <InputContainer locked={true}>
               <input
                 disabled
                 type="text"
-                id={nomination.nominationID}
+                id="nomination"
                 value={nomination.electionSymbol ? null : $t('candidateApp.basicInfo.pending')}
                 class={inputClass} />
-              <Icon name="locked" class={iconClass} />
-            </div>
+            </InputContainer>
           </Field>
-        {/each}
+        {/if}
+
         <p class={disclaimerClass} slot="footer">
           {$t('candidateApp.basicInfo.nominationsDescription')}
         </p>
@@ -235,9 +253,10 @@
           <label for="birthday" class={labelClass}>
             {$t('candidateApp.basicInfo.fields.birthday')}
           </label>
-          <div class={inputContainerClass}>
+          <InputContainer locked={questionsLocked}>
             <div class={inputClass}>
               <input
+                disabled={questionsLocked}
                 class="dark:bg-black"
                 type="date"
                 min={birthdayMin}
@@ -245,25 +264,27 @@
                 id="birthday"
                 bind:value={birthday} />
             </div>
-          </div>
+          </InputContainer>
         </Field>
         <Field>
           <label for="gender" class={labelClass}>
             {$t('candidateApp.basicInfo.fields.gender')}
           </label>
-
-          <select
-            id="gender"
-            class={selectClass}
-            bind:value={genderID}
-            style="text-align-last: right; direction: rtl;">
-            {#if allGenders}
-              {#each allGenders as option}
-                <option value={option.id} selected={option.id === genderID}
-                  >{$t(`candidateApp.genders.${option.attributes.name}`)}</option>
-              {/each}
-            {/if}
-          </select>
+          <InputContainer locked={questionsLocked}>
+            <select
+              disabled={questionsLocked}
+              id="gender"
+              class={selectClass}
+              bind:value={genderID}
+              style="text-align-last: right; direction: rtl;">
+              {#if allGenders}
+                {#each allGenders as option}
+                  <option value={option.id} selected={option.id === genderID}
+                    >{$t(`candidateApp.genders.${option.attributes.name}`)}</option>
+                {/each}
+              {/if}
+            </select>
+          </InputContainer>
         </Field>
       </FieldGroup>
 
@@ -271,47 +292,56 @@
         <p class={headerClass} slot="header">
           {$t('candidateApp.basicInfo.fields.motherTongue')}
         </p>
-
-        <Field>
-          <label for="motherTongue" class={labelClass}>
-            {#if motherTongues}
-              {motherTongues.length > 0
-                ? $t('candidateApp.basicInfo.addAnother')
-                : $t('candidateApp.basicInfo.selectFirst')}
-            {/if}
-          </label>
-          <select
-            bind:this={motherTongueSelect}
-            id="motherTongue"
-            class={selectClass}
-            on:change={handleLanguageSelect}>
-            <option disabled selected value style="display: none;" />
-            {#each availableLanguages ?? [] as option}
-              <option value={option.attributes.localisationCode}
-                >{$t(`candidateApp.languages.${option.attributes.name}`)}</option>
-            {/each}
-          </select>
-        </Field>
+        {#if !questionsLocked}
+          <Field>
+            <label for="motherTongue" class={labelClass}>
+              {#if motherTongues}
+                {motherTongues.length > 0
+                  ? $t('candidateApp.basicInfo.addAnother')
+                  : $t('candidateApp.basicInfo.selectFirst')}
+              {/if}
+            </label>
+            <InputContainer locked={questionsLocked}>
+              <select
+                disabled={questionsLocked}
+                bind:this={motherTongueSelect}
+                id="motherTongue"
+                class={selectClass}
+                on:change={handleLanguageSelect}>
+                <option disabled selected value style="display: none;" />
+                {#each availableLanguages ?? [] as option}
+                  <option value={option.attributes.localisationCode}
+                    >{$t(`candidateApp.languages.${option.attributes.name}`)}</option>
+                {/each}
+              </select>
+            </InputContainer>
+          </Field>
+        {/if}
         {#each motherTongues ?? [] as tongue}
           <Field>
             <label for={tongue.name} class={labelClass}>
               {tongue.name}
             </label>
             <div class={buttonContainerClass}>
-              <button
-                title="remove"
-                type="button"
-                id={tongue.name}
-                on:click={() => (motherTongues = motherTongues?.filter((m) => m.id !== tongue.id))}>
-                <Icon name="close" class={iconClass} />
-              </button>
+              {#if !questionsLocked}
+                <button
+                  title="remove"
+                  type="button"
+                  id={tongue.name}
+                  on:click={() =>
+                    (motherTongues = motherTongues?.filter((m) => m.id !== tongue.id))}>
+                  <Icon name="close" class={iconClass} />
+                </button>
+              {:else}
+                <Icon name="locked" class={iconClass} />
+              {/if}
             </div>
           </Field>
         {/each}
       </FieldGroup>
 
       <FieldGroup>
-        <AvatarSelect bind:photo bind:uploadPhoto />
+        <AvatarSelect bind:photo bind:uploadPhoto disabled={questionsLocked} />
       </FieldGroup>
 
       <FieldGroup>
@@ -319,12 +349,17 @@
           <label for="unaffiliated" class={labelClass}>
             {$t('candidateApp.basicInfo.fields.unaffiliated')}
           </label>
-
-          <input
-            id="unaffiliated"
-            type="checkbox"
-            class="toggle toggle-primary mr-8"
-            bind:checked={unaffiliated} />
+          <InputContainer locked={questionsLocked}>
+            {#if !questionsLocked}
+              <input
+                id="unaffiliated"
+                type="checkbox"
+                class="toggle toggle-primary mr-8"
+                bind:checked={unaffiliated} />
+            {:else}
+              <input id="unaffiliated" disabled value="yes" class={inputClass} />
+            {/if}
+          </InputContainer>
         </Field>
         <p class={disclaimerClass} slot="footer">
           {$t('candidateApp.basicInfo.unaffiliatedDescription')}
@@ -332,6 +367,7 @@
       </FieldGroup>
 
       <MultilangTextInput
+        locked={questionsLocked}
         id="manifesto"
         headerText={$t('candidateApp.basicInfo.electionManifesto')}
         localStorageId="candidate-app-manifesto"
@@ -342,11 +378,20 @@
 
       <Button
         disabled={!allFilled || loading}
-        text={$t('candidateApp.questions.continue')}
+        text={submitButtonText}
         type="submit"
         variant="main"
         icon="next"
         slot="primaryActions" />
+      {#if opinionQuestionsFilled && !questionsLocked}
+        <Button
+          color="error"
+          on:click={async (event) => {
+            event.preventDefault();
+            await goto($getRoute(Route.CandAppHome));
+          }}
+          text={$t('candidateApp.navbar.cancel')} />
+      {/if}
       {#if errorMessage}
         <div class="text-error">
           {errorMessage}
